rename modules according to review
