+++ core/eolearn/core/__init__.py
@@ -28,8 +28,8 @@ from .core_tasks import (
     MergeEOPatchesTask,
 )
 
-from .utilities.filesystem import get_filesystem, load_s3_filesystem
-from .utilities.feature_parser import FeatureParser
-from .utilities.other import deep_eq, constant_pad
+from .utils.fs import get_filesystem, load_s3_filesystem
+from .utils.parsing import FeatureParser
+from .utils.common import deep_eq, constant_pad
 
 __version__ = "1.0.0"
