+++ core/eolearn/core/utils/parsing.py
